+++ src/service/mod.rs
@@ -1,3 +1,4 @@
 pub mod build_system;
+pub mod config;
 pub mod init;
 pub mod vcs;